backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
redis
